// pi-nail-server/src/models/IConfig.ts
/*
 * File: c:\pi-nail\pi-nail-server\src\models\IConfig.ts
 * Project: c:\pi-nail\pi-nail-server
 * Created Date: Sunday April 15th 2018
 * Author: J-Cat
 * -----
 * Last Modified:
 * Modified By:
 * -----
 * License: 
 *    This work is licensed under a Creative Commons Attribution-NonCommercial 4.0 
 *    International License (http://creativecommons.org/licenses/by-nc/4.0/).
 * -----
 * Copyright (c) 2018
 */
export interface IConfig {
    p: number;
    i: number;
    d: number;
    setPoint: number;
    maxPower: number;
    maxTemp: number;
    tcInterval: number;
    cycleTime: number;
}